requirements.txt
# conda create -n myenv python=3.10
# conda activate myenv
# pip install -r requirements.txt 
h5py
numpy
matplotlib
pillow
json
nibabel
loguru
pathlib